backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
fsrs==6.1.0
